LPS-89787 Move it to a v1_0 dir
